New feature multiple selection was added.
New feature to select multiple files with the option multiplePicker, async task was changed by coroutine worker, this is the new reccommended way by google to make async task.

Some code refactor.

Commented  // mCropProvider.delete() because, it was removing image after generate, it was causing an error that image wasn't available when client wanted to consumed.